Create queues module
